pyproject.toml
[project]
name = "kinesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.2",
    "pillow>=12.1.1",
]
